Vision cone collider trigger working
This was a nightmare, I originally wanted to use a 2D triangle sprite, with a 2d polygon collider - but this collider would not let me edit it, it kept staying as a line, something was wrong with the quaternions, as I had to rotate the sprite (which I don't want since it looks weird), and I noticed the 2d polygon collider was having 4d reflections, suuuuper weird stuff.

Reverted to using a simple 3d cone with a mesh collider - not as clean as I'd like but it works.