Updating double click algorithm
Had some real issues getting this one working.
My double click managers were fine.

I removed all the redundant ones I tried working on, and am now only using "ClickEventListener"

The issue was that the raycast was checking for hits with the 'floor' tag, the raycast was hitting the 'click indicator' visual, and thus not responding to the double click as it was not the floor. Had to set the 'click indicator' to ignore raycast layer.